requirements.txt
Flask==2.3.2
pymongo==4.4.0
redis==4.5.5
cassandra-driver==3.25.0
neo4j==5.14.0
python-dotenv==1.0.0
